split profile components
