+++ components/InitialProfile/index.ts
@@ -1,0 +1,1 @@
+export { InitialProfileImage } from './InitialProfileImage';
